// src/app/models/app-template-conversion-unit-update-dto.ts
export class AppTemplateConversionUnitUpdateDto {

    id: number;
    appUnitIdSince: number;
    appUnitIdUntil: number;
    appVariableId: number;
    description: string;
    value: number;
    formula: string;
    formulaValue: string;
    sumValue: boolean;
    orderCalculate: number;
    code: string;

}