// src/app/models/app-template-conversion-unit-create-dto.ts
export class AppTemplateConversionUnitCreateDto {


    appUnitIdSince: number;
    appUnitIdUntil: number;
    appVariableId: number;
    description: string;
    value: number;
    formula: string;
    formulaValue: string;
    sumValue: boolean;
    orderCalculate: number;
    code: string;

}